requirements.txt
Pillow==8.1.0
streamlit==0.78.0
opencv-contrib-python==4.5.1.48
numpy==1.19.5
opencv-python==4.5.1.48